# healy-apigateway/cmd/docs/swagger.yaml
definitions:
  models.AdminLogin:
    properties:
      email:
        type: string
      password:
        maxLength: 20
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  models.AdminSignUp:
    properties:
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      password:
        maxLength: 20
        minLength: 6
        type: string
    required:
    - email
    - firstname
    - lastname
    - password
    type: object
...
  models.GoogleUserInfo:
    properties:
      accesstoken:
        type: string
      email:
        type: string
      id:
        description: Google's unique user ID
        type: string
      name:
        type: string
      refreshtoken:
        type: string
      tokenexpiry:
        type: string
    type: object
  models.PatientDetails:
    properties:
      contactnumber:
        description: |-
          The contact number of the patient
          example: +1234567890
        type: string
      email:
        description: |-
          The email address of the patient
          example: john.doe@example.com
        type: string
      fullname:
        description: |-
          The full name of the patient
          example: John Doe
        type: string
      gender:
        description: |-
          The gender of the patient
          example: male
        type: string
    type: object
...
  response.Response:
    properties:
      data:
        description: |-
          The response data
          example: {"key": "value"}
      error:
        description: |-
          The error information, if any
          example: "Error description"
      message:
        description: |-
          The response message
          example: "Request was successful"
        type: string
    type: object
info:
  contact:
    email: support@healy-medicare.com
    name: API Support
  description: This is the API documentation for the Healy Medicare application.
  title: Healy Medicare API
  version: 1.0.0
paths:
...
  /admin/dashboard/patients:
    get:
      description: List all patients
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: List Patients
      tags:
      - Admin
  /admin/login:
    post:
      description: Admin Login endpoint
      parameters:
      - description: Admin Login
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/models.AdminLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin Login
      tags:
      - Admin
  /admin/signup:
    post:
      description: Admin SignUp endpoint
      parameters:
      - description: Admin SignUp
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/models.AdminSignUp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin SignUp
      tags:
      - Admin
...
  /google/redirect:
    get:
      consumes:
      - application/json
      description: Handles the callback from Google OAuth2 login
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.GoogleUserInfo'
        "400":
          description: No code in query parameters
          schema:
            type: string
        "500":
          description: Failed to exchange token
          schema:
            type: string
      security:
      - None: []
      summary: Handle Google OAuth2 callback //redirect from google login
      tags:
      - Patients
...
  /patient/login:
    get:
      consumes:
      - application/json
      description: Redirects the user to Google's OAuth2 login page
      produces:
      - application/json
      responses:
        "302":
          description: Redirecting to Google login
          schema:
            type: string
      security:
      - None: []
      summary: Redirect to Google OAuth2 login //use browser
      tags:
      - Patients
  /patient/profile:
    get:
      consumes:
      - application/json
      description: Retrieve details of a specific patient
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get patient details
      tags:
      - Patient Profile
    put:
      consumes:
      - application/json
      description: Update the details of a patient
      parameters:
      - description: Patient details to update
        in: body
        name: patientDetails
        required: true
        schema:
          $ref: '#/definitions/models.PatientDetails'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update patient details
      tags:
      - Patient Profile
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"